Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11007 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: b14be8daffd5fd70f6fb88d7dd3cd2772dc11f0c

+++ costmodel/de.uniol.inf.is.odysseus.costmodel.operator.datasrc/src/de/uniol/inf/is/odysseus/costmodel/operator/datasrc/impl/AttributeObserver.java
@@ -69,7 +69,7 @@ public class AttributeObserver implements IDataSourceObserverListener {
 		// direct input... e.g. cached values
 		if( element instanceof Double ) {
 			factory.addValue((Double)element);
-			getLogger().debug("Direct value for " + attribute + ": " + ((Double)element) );
+			getLogger().debug("Direct value for " + attribute + ": " + element );
 			return;
 		}
